Purge old read caching code.
