+++ virtool_workflow/analysis/utils.py
@@ -1,3 +1,4 @@
+import shutil
 from pathlib import Path
 from typing import Tuple, Callable, Union
 
@@ -9,8 +10,8 @@ def _make_paired_paths(
         paired: bool,
         mkstr: Callable[[int], str]
 ) -> ReadPaths:
-    path1 = dir_path/mkstr(1)
-    return (path1, dir_path/mkstr(2)) if paired else (path1,)
+    path1 = dir_path / mkstr(1)
+    return (path1, dir_path / mkstr(2)) if paired else (path1,)
 
 
 def make_read_paths(
@@ -33,3 +34,30 @@ def make_legacy_read_paths(
 ) -> ReadPaths:
     return _make_paired_paths(reads_dir_path, paired, lambda n: f"reads_{n}.fastq")
 
+
+def rename_trimming_results(path: Path):
+    """
+    Rename Skewer output to a simple name used in Virtool.
+
+    :param path: The path containing the results from Skewer
+    """
+    try:
+        shutil.move(
+            path / "reads-trimmed.fastq.gz",
+            path / "reads_1.fq.gz",
+        )
+    except FileNotFoundError:
+        shutil.move(
+            path / "reads-trimmed-pair1.fastq.gz",
+            path / "reads_1.fq.gz",
+        )
+
+        shutil.move(
+            path / "reads-trimmed-pair2.fastq.gz",
+            path / "reads_2.fq.gz",
+        )
+
+    shutil.move(
+        path / "reads-trimmed.log",
+        path / "trim.log",
+    )
